pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "offscroll"
version = "0.1.0"
description = "Your personal newspaper from the open web"
readme = "README.md"
license = {text = "Apache-2.0"}
requires-python = ">=3.11"
authors = [
    {name = "Modus Ponens", email = "modusponies@pm.me"},
]
keywords = ["newspaper", "rss", "curation", "self-hosted", "pdf"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Communications :: Email",
    "Topic :: Internet :: WWW/HTTP :: Dynamic Content :: News/Diary",
    "Topic :: Printing",
    "Topic :: Text Processing :: Markup :: HTML",
]
dependencies = [
    "click>=8.1",
    "pyyaml>=6.0",
    "feedparser>=6.0",
    "httpx>=0.27",
    "weasyprint>=62",
    "Jinja2>=3.1",
    "scikit-learn>=1.4",
    "numpy>=1.26",
    "hdbscan>=0.8",
    "ollama>=0.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.7",
    "mypy>=1.11",
]
ollama = []
cloud = [
    "anthropic>=0.39",
    "openai>=1.50",
]
fediverse = [
    "Mastodon.py>=1.8",
    "atproto>=0.0.50",
]

[project.urls]
Homepage = "https://github.com/offscroll/offscroll"
Repository = "https://github.com/offscroll/offscroll"
Issues = "https://github.com/offscroll/offscroll/issues"

[project.scripts]
offscroll = "offscroll.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "slow: marks tests that call external APIs or take >5s",
]
